ensure we reset contexts when fixture is used
Previously using the ClearRequestContext fixture only reset contexts
on fixture teardown. If *all* tests in a test suite use this fixture,
this is fine. However if any tests do not, it means you might start
with a carry over context from previous tests.

Using this fixture means you really wanted a clean slate, so fix it so
that you get a clean state when you get started as well as when you
finish.

Change-Id: I5e6f7bc2866fe66f269c7d5e1f36c711151489c2
Closes-Bug: #1428728