Fix context from_dict() for system_scope
A previous patch[1] added "system_scope" to the context for to_dict()
but the from_dict() method was not updated. This caused the
system_scope to always be None.
This patch corrects that by adding "system_scope" to the list of
values that may need to be extracted from the context dict.

[1] https://review.opendev.org/c/openstack/oslo.context/+/530509

Change-Id: Ica23d5c4183a692de3cb65a7ad72b19f47988ca6