Don't test with setuptools local distutils.
We noticed some aberrations in log level/threshold handling as of
Setuptools 60.0.0, and the only substantive change for that release[1]
toggles the default selection for its internal ("local") distutils
replacement vs the "stdlib" one. For now, toggle the
SETUPTOOLS_USE_DISTUTILS envvar in our testing back to the old
value. This should be revisited once logging is handled better in
Setuptools.

[1] https://github.com/pypa/setuptools/issues/3038

Change-Id: I2eacb9756d9eb9c5d606599f16cdb412007f5854