Add is_admin to redacted context
Heat is known to override RequestContext in a strange way and it
attempts to run policy check, which loads full request context, in case
is_admin is not given.

This ensures is_admin is copied during redactation, to avoid failures
in unit tests. is_admin is not a secret field so can appear in
notifications.

This also adds **kwargs interface to redacted_copy, so that we can
inject a few more arguments in subclasses.

Change-Id: I19f83d7b0a13f14df16ac805998d5b3751801c12