Add build-system to pyproject.toml
pip 23.1 removed the "setup.py install" fallback for projects
that do not have pyproject.toml and now uses a pyproject.toml
which is vendored in pip.
To address that, this change adds the minimal pyproject.toml
to enable pbr to be properly used to build editable wheels.

See https://pip.pypa.io/en/stable/news/#v23-1
and https://github.com/pypa/pip/issues/8368 for more
details on the removal of the fallback support.

Co-Authored-By: Sean Mooney <work@seanmooney.info>
Change-Id: I93357940617a416dc684e9228ea6747d91b62c44